apps/backend/requirements.txt
torch
httpx
pytest
fastapi
uvicorn
faiss-cpu
python-multipart
sentence-transformers
